Specify the charset as UTF-8.
Since Testling didn't specify the charset, this led to a plethora of random encoding errors if the libraries used anything outside of the ASCII charset. The same code would work or not on browsers with the same capabilities depending on whether the browser decided to treat the page as UTF-8 or not.

Specifying the charset gets rid of all that and gives us predictable behaviours. Yay